Filepond disk tests added ✅
