+++ tests/TestCase.php
@@ -26,7 +26,7 @@ class TestCase extends \Orchestra\Testbench\TestCase
         include_once __DIR__.'/../database/migrations/create_fileponds_table.php.stub';
         include_once __DIR__.'/database/migrations/create_users_table.php.stub';
 
-        (new \CreateUsersTable())->up();
-        (new \CreateFilepondsTable())->up();
+        (new \CreateUsersTable)->up();
+        (new \CreateFilepondsTable)->up();
     }
 }
